Minor refactoring and min/max properties on sets
